Factor out `mirror_path` into path module
